requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-decouple>=3.8
openai>=1.12.0
Pillow>=10.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0

